Use Apache Fileupload streaming API
The Apache Commons Fileupload library offers two different APIs, a
traditional one that stores uploads in temp files before making them
available, and a streaming API that provides direct access to the
upload stream.

The streaming API has many benefits for this add-on:
* it fits better with the concept of a receiver that handles uploaded
  contents
* it makes this add-on independent from the filesystem
* it does not require cleaning up temporary files, we can get rid of
  the CleaningTracker
* it's faster because we don't need to copy streams twice

As a side effect of this change, it's not directly possible to support
the method FileDetails.getContentLength() anymore. However, returning
-1 for "unknown" is in conformance with the contract of this method.